uklid prav a cc na uzivatelpresenteru
