+++ app/templates/UzivatelRightsCc/editrights.latte
